Extract some common code in tests/common
