+++ tests/test_lang.py
@@ -1,13 +1,11 @@
 
 from copy import copy
-from types import SimpleNamespace
-from dataclasses import dataclass
-
 from pytest import mark
+from types import SimpleNamespace
 
 from myia.api import standard_debug_pipeline
-from myia.dtype import Int
-from myia.ir import MultitypeGraph
+
+from .common import mysum, Point
 
 
 lang_pipeline = standard_debug_pipeline \
@@ -464,25 +462,6 @@ def test_rec1(x):
 #############
 
 
-mysum = MultitypeGraph('mysum')
-i64 = Int[64]
-
-
-@mysum.register(i64)
-def _mysum1(x):
-    return x
-
-
-@mysum.register(i64, i64)
-def _mysum2(x, y):
-    return x + y
-
-
-@mysum.register(i64, i64, i64)
-def _mysum3(x, y, z):
-    return x + y + z
-
-
 @parse_compare((2, 3, 4))
 def test_multitype(x, y, z):
     return mysum(x) * mysum(x, y) * mysum(x, y, z)
@@ -523,15 +502,6 @@ def test_fact(x):
         else:
             return n * fact(n - 1)
     return fact(x)
-
-
-@dataclass(frozen=True)
-class Point:
-    x: Int[64]
-    y: Int[64]
-
-    def abs(self):
-        return (self.x ** 2 + self.y ** 2) ** 0.5
 
 
 @parse_compare(42)
